refactor: module-domain 내부 image 도메인 패키지 이동 (#107)
